[#185] Move relative to absolute to file open
