+++ test/coverage-system/topSectionFinder.test.ts
@@ -1,6 +1,8 @@
 import {Section} from "lcov-parse";
 import {TextEditor} from "vscode";
+
 import {SectionFinder} from "../../src/coverage-system/sectionFinder";
+import {fakeConfig} from "../mocks/fakeConfig";
 
 suite("SectionFinder Tests", function() {
 
@@ -21,7 +23,7 @@ suite("SectionFinder Tests", function() {
     test("Should not throw an error @unit", function(done) {
         const textEditor: TextEditor = {} as TextEditor;
         const sectionMap: Map<string, Section> = new Map<string, Section>();
-        const sectionFinder: SectionFinder = new SectionFinder(fakeOutput, fakeReporter as any);
+        const sectionFinder: SectionFinder = new SectionFinder(fakeConfig, fakeOutput, fakeReporter as any);
         sectionFinder.findSectionForEditor(textEditor, sectionMap);
         return done();
     });
